ENH: Added the Q1 hack
Instead of evaluating the Rumpf functional for the Q1 trafo, each
hypercube is divided into simplices and there the functional for the P1
trafo is evaluated. The crucial point is that this has to be done for
all permutations, meaning for every possibility of dividing a hypercube
into simplices.

Another point is that h has to be different in norm_A and det_A so that
the identity mapping on the reference element results in the minimal
of functional value over all admissible deformations.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>