ENH: Added LaplaceSmoother
Assume that for a given mesh, the boundary is moved and that the
interiour vertices should be adjusted according to this. The
LaplaceSmoother minimises the harmonic energy by solving a discrete
Laplace problem on the undeformed mesh on the FE space defined by the
transformation, with the new vertex coordinates at the boundary as
Dirichlet boundary conditions.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>