ENH: revised Foundation::HaloFrequencies implementation
The previous implementation was more or less hard-coded for
LAFEM::DenseVector containters, so it did not work for meta-vectors. The
new implementation utilises only LAFEM operations, so it now works for
meta-vectors as well.

As a side-effect, the HaloFrequencies function now only uses a single
frequency buffer vector instead of a whole set of those. All codes using
the HaloFrequencies class have been adapted.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>