ENH: added BandedBilinearOperatorTest
For testing assembly of the SparseMatrixBanded a new test has been added.
This new test works on a StructuredMesh instead of a ConformalMesh, which
now really leads to a banded matrix rather than an almost dense one stored
in banded format. In consequence, the SparseMatrixBanded has been removed
from the BilinearOperatorTest working on structured meshes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>