ENH: Added apply1 and apply2 routines
These apply a bilinear operator to an FE function and every test
function. This is the same as assembling the bilinear operator into a
matrix and performing one matrix vector product. If this is needed only
a handful of time (i.e. in assembling the rhs for a time
discretisation), using these routines is more efficient than the matrix
assembly.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>