ENH: added element regression test
This new test serves as a complex "test battery", whose purpose is to
ensure that future changes in element implementations do not yield
different results. In contrast to the other element tests this test
battery does not ensure that the element implementation is actually
correct, but only to detect if changes have any unexpected effects.

The reference error, which are hard-coded in the test source file, are
computed by test runs and are manually certified as being 'most probably
correct' by some expert in finite element stuff.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>