SYS: Added build system informations to feast_config.hpp
feast_config.hpp should contain all information to reproduce the
configuration and build process of a given FEAST checkout.

The only known exception of this assumption is the usage of compiler
wrappers. In this case, the actual host compiler is choosen at runtime
and therefore its path and version may differ from the data in
feast_config.hpp.
You may rely on FEAST_COMPILER, which is only present at 'compile time'
and thus cannot be included in feast_config.hpp.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>