ENH: added andicore-mg tutorial
This 'tutorial' has already been used as a code-base for various bachelor
and master thesises. Since we assume that even more work based on this
code will follow, the decision was made to inlude this code in the main
development branch.

Please note that the main purpose of this application is of dydactic
nature, i.e. it is more a code-base to start off rather than a tutorial to
learn from, therefore it is labeled as 'Tutorial #X1' rather than
'Tutorial #04'.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>