fixed some flaws in Random and added RandomTest
Turns out that even a random number generator, whose only purpose is to
produce junk, can produce junk junk. More precisely: the ranged evaluation
operator, which returns a random number in a range [a,b], did not work as
expected for signed integer types. Therefore, a test class for the RNG has
been added to ensure that the junk is not junk.
Moreover, the implementation to extract a 'bool' from the rng has been
added. Finally, some minor syntactical and semantical enhancements have
been incorporated.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>