enhanced IterativeSolver base class
The provisional IterativeSolver base class has been enhanced with a few
new features:

The class now also has an absolute tolerance parameter in addition to
the already existing relative tolerance parameter. An iterative solution
process is treated as 'converged' if both the absolute and relative
tolerances are fulfilled. Howver, by default the absolute tolerance is set
to 1/eps^2 (~1E+32 for double), which effectively disables the absolute
tolerance.

Moreover, in analogy to the tolerances, relative and absolute divergence
criterions have been added. An iterative solution process is treated as
'diverged' if one of the absolute or relative criterions is fulfilled.

Finally, a few 'getter' functions for most of the interesting member
variables have been added, including a function which returns the
convergence rate for the iterative solver based on the defects.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>