added LaplaceFunctional and ExpBubbleFunction
This commit adds a new non-polynomial analytic function to the set of
common functions, as well as the strong Laplace functional, which can
be used for the right-hand-side vector assembly of a Poisson system for
any given analytic solution function.
As a small addition, the Tiny::Matrix can compute its trace now.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>