minor UnitFilter(Blocked) updates
The UnitFilter and UnitFilterBlocked class templates had a size()
member function which returned the number of non-zero entries of
the underlying SparseVector(Blocked) container. This was kind of
irritating, since this did not correspond to the total (virtual)
size of the underlying container.

In consequence, the UnitFilter(Blocked) classes now have a member
function 'used_elements()', which returns the number of non-zeros
in the SparseVector(Blocked), as well as a member function
'size()', which returns the total size of the underlying container.

This means that the semantic of the 'size()' member function has
changed, however, it seems that the previous implementation was
never used, so I do not expect any side effects here.

Finally, the UnitFilter(Blocked) classes now also have a new
member function named 'get_filter_vector()', which returns a
(const) reference to the underyling SparseVector(Blocked).

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>