ENH: Added FEniCS code for mesh smoothers
For toying around with different mesh smoothers, the supplied code
offers some basic tools. There is even a true FE implementation of the
RumpfSmoother in several variations. The r-adaptivity does not work,
though, because there is simply no solver that can solve the resulting
highly nonlinear system.

There is one workaround for setting the boundary values for the
deformation due to a possible bug in FEniCS 1.4. With this, it now is
possible to specify time dependant BCs that are nonlinear in time.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>