Removed Algo Tag from all FEAST operations
The used algorithm backend is now automatically choosen depending on the
build id and availability of operation.
This reduces the amount of template parameters and enables the 'mixing'
of operations with and without third library support.

The algo classes still remain in archs.hpp for benchmarking purposes.

Adapting your own branch after a rebase:
* Remove any Algo_ template parameters from your classes.
* Remove Algo_ or Algo::Generic parameters from all lafem operation
  calls.
  This usually includes the 'template' keyword, related to the
  operation call, because the lafem operations are not template functions any
  more.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>