enhanced TimeStamp and StopWatch classes
The TimeStamp class' functionality of writing out the elapsed time as a
formatted string has been outsourced into a static function named
'format_micros()'. Moreover, this new function and the elapsed_string
function have additional formatting parameters which specify whether hours
and milliseconds are to be printed.

The StopWatch class has been enhanced in two ways:
First, now keeps track of whether it is currently running or not. If the
stop watch is running, the 'elapsed_*' functions returns the current
running time. If the stop watch is not running, the 'elapsed_*' functions
return the time that elapsed while the watch was running.
In other words, the StopWatch class now behaves like a real-life
stop-watch.

Furthermore, the StopWatch class also supports a 'elapsed_string()'
function, which returns the elapsed time as a formatted string in analogy
to the TimeStamp class.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>