enhanced SimpleArgParser class
This commit introduces a few changes to the SimpleArgParser class.

First, the 'query' function, which is used to parse the parameters of a
particular option, has been renamed to 'parse'.

Second, a new 'query' function has been added, which returns the deque of
all parameters for a particular option. This function can be used if one
wants to parse parameters, which blow the capability of the 'parse'
function.

Moreover, it turned out that using command line arguments is quite error
prone if the parser does not check for unsupported options (think of
typos). Therefore, a new 'support' function has been added, which should
be called for each supported option to specify the set of all options
supported by the application. Afterwards, one can use the new
'query_unsupported' function to check whether any unsupported options have
been given in the command line. See the documentation of the
SimpleArgParser class for an example.

Finally, some internal code restruction was performed.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>