added inner FGMRES(k) residual scale factor
The previous implementation always performed the full inner GMRES(k) loop
independently of whether the maximum number of iterations was performed or
whether the inner residual achieved the desired tolerance.
Now one can provide a scaling factor for the tolerance of the inner
residual. If this tolerance is achieved, the inner FGMRES(k) loop breaks
and a restart is performed.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>