added matrix-cond tool
This new tool is capable of computing the maximal singular values and the
spectral condition number of a given matrix. The SVs are computed by
applying the power method onto the normal system, which can be configured
by various options.

Note that this tool is still experimental, so use with care.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>