ENH: MeshStreamerFactory can now handle meshes with incomplete information
The original design was that mesh files have to provide complete
adjacency information, like vertex@edge, vertex@face arrays. This allows
for boundary identification and uniquely determines the complete mesh
adjacency structure.

However, the only critical information is the vertex@cell array and the
list of vertices. The other arrays can be computed from this
information, although care has to be taken in the MeshStreamerFactory.
In these minimal meshes, no boundary identification is possible, so if
the user wants to save some work by not providing this information in
the mesh file itself, he has to manually identify different boundaries
i.e. by using the BoundaryFactory and problem specific code.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>