implemented initialisation for VC test system
Note: This commit only affects the Visual Studio test system for Windows
and has no effect on other platforms.

The default behaviour on a Windows application crash is to display a
message box dialog telling the user what happened, offering the
possibility to attach the debugger for solving the root of the cause.
This has already been disabled for the Visual Studio test system a long
time ago, since dialog boxes waiting for user interaction are somewhat
counter-productive in an automated regression test system. The drawback
was that test crashes did not produce any type of output to stdout/stderr,
so that no information about the crash were written to the test log files,
which are commited to the dash site.

This commit adds a 'structed exception handling' filter, which catches any
type of error left unhandled by the test application (e.g. segfaults) and
writes out some information to stderr, including a call-stack back-trace.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>