revised VTK exporter to support partitioned meshes
The (previously provisional) VTK exporter has been revised to export to
the 'new' XML-based VTK file format instead of the 'old' legacy format.
The primary reason is to support the export of partitioned meshes into the
VTU/PVTU file formats.

Most of the changes are either new features (not breaking backwards
compatibility) or changes performed 'under the hood'.

However, there is one notable change for users:

When calling the (one-argument) 'write' method of the ExportVTK
object, one does not specify the file extension anymore. The file
extension, which is '.vtu' for stand-alone meshes, is appended
automatically by the exporter.

In addition, there is a new 'write' method, which takes the rank and
number of partitions/processes as arguments in addition to the filename.
Note that this function is also available in non-MPI builds despite the
need for a rank and a partition count. This function performs two tasks:

First, it calls the standard 'write' function to write out the actual mesh
partition data into a VTU file, where the filename is postfixed by the rank,
i.e. the filename will have the form '<filename>.<rank>.vtu'.

Second, if the rank passed to the function is 0, the function will also
write the corresponding 'parallel' PVTU file containing the list of all
VTU files that make up the whole partitioned mesh. The filename of the
PVTU file is of the form '<filename>.pvtu'. (BTW, that's the reason why
the extension is not specified explicitly by the user anymore.)

Final Note: The ExportVTK class template supports both ConformalMesh and
StructuredMesh objects, however, the output format is always an
unstructured mesh in the sense of VTK.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>