added empty() member function to LAFEM::Container
In analogy to the STL containers, this function returns true if the
container is empty and false otherwise. Moreover, this function was used
to fix the VectorMirror, which crashed when the underlying mirror matrix
was empty. Note that an empty mirror is indeed a valid case as this is the
natural and desired result when using discontinuous FE spaces.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>