added initialise, finalise and abort functions
This commit adds three new global functions:

* initialise()
This function performs some basic static initialisation and should be
called as the first statement in the main() function of any application.
This function has two input arguments, namely the argc and argv parameters
of the calling main function, and it has two (optional) output arguments,
which receive the rank and the number of processes in a parallel run.
If MPI is enabled, this function takes care of initialising MPI as well as
retrieving the current rank and process count.

* finalise()
This function performs the basic cleanup and should be called as the last
statement before the return of the main() function of any application.
This function returns the constant 'EXIT_SUCCESS', so that one may simply
write 'return finalise();' as the last statement of main.
If MPI is enabled, this function takes care of finalising MPI.

* abort()
This function aborts the current process as well as all other processes in
an MPI run. This function is effectively just a wrapper around MPI_Abort
and std::abort(), resp.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>