added first four parallel applications
This commit introduces global matrix, vector and filter container wrappers
located in the FEAST::Global namespace. These wrappers are used for the
implementation of MPI-based global linear systems, which are solved in
the four new parallel applications.

Note that these applications are still in a prototypic state. Their main
purpose is to serve as a testing and development framework for further
kernel code restructuring.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>