Revised the time stamp string formating
Introduced the enum TimeFormat, controlling the actual string formating.

For TimeFormat::h_m_s_m the string is of the format h:mm:ss.mmm.
For TimeFormat::h_m_s the string is of the format h:mm:ss.
For TimeFormat::m_s_m the string is of the format mm:ss.mmm.
For TimeFormat::m_s the string is of the format mm:ss.
For TimeFormat::s_m the string is of the format ss.immm.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>