ENH: Added MeshWriter
With the MeshWriter (that resides in mesh_streamer_factory.hpp for the
time being) it is possible to write Geometry::MeshNodes to FEAT mesh
files. The complete tree is written, including all MeshParts and Charts.

Charts now need to provide their own method for writing their data. This
is not very pretty but the least invasive method of doing so, as the
data varies greatly and only the Chart itself knows who it is (and
possibly it knows its father as well).

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>