fixed harmlessly looking but actually quite serious bug
The VC14 compiler kept complaining about unused variables in opt build
mode. After closer investigation, the faulty lines were something like:

<code>
  for(auto m : meshes)
    ASSERT(...)
  blabla();
  ...
</code>

Now, in opt build mode, ASSERT evaluates to an empty macro, so the code
actually does *not* produce an empty loop, but instead this leads to

<code>
  for(auto m : meshes)
    blabla();
  ...
</code>

which is certainly not what it was meant to be.

Fixed this issue by placing an #ifdef DEBUG around the three affected
loops.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>