ENH: Added SlipFilter
Enforcing slip boundary conditions is easy for isoparametric conforming
Lagrange elements because it can be done in the strong sense. For other
cases, it is not so clear what kind of approximate condition can be
enforced, so the only assembly of a SlipFilter is for conforming
Lagrange elements of order 1 and the standard trafo.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>