implemented Vanka preconditioner/smoother
This commit introduces an implementation of the Vanka, which is a special
preconditioner/smoother for saddle-point-systems.

This new class implements a total of eight different variants, which are
the cross-product of three binary variants:

 1. Blocked vs Nodal variants
 2. Diagonal vs Full variants
 3. Additive vs Multiplicative variants

Note that the desired variant can be chosen at runtime.

Currently, the Vanka implementation does not support global (i.e. type-0)
matrices - this has yet to be figured out...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>