ENH: Use FilterChain in mesh optimisers
For mesh optimisation problems, useful boundary conditions are Dirichlet
and slip boundary conditions. Mesh optimisers now contain a FilterChain
that first applies slip BCs, then Dirichlet BCs.

One nonsensical exception is the LaplaceSmoother, which treats the
components as decoupled. This is only the case if the slip boundaries
are aligned with the coordinate axis. There might be a way to figure
this out later, but at the moment, everything slip is ignored in the
LaplaceSmoother.

Because the boundary treatment got more complex, the MeshSmoother now
contains a pointer to the RootMeshNode (instead of just the mesh).

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>