Changed filter assembly to up/download and assemble
Before, the UnitFilterAssembler and SlipFilterAssember used
filter.add(...) which under the hood used the SparseVector ()-operator
to add entries to the filter. These operations are highly inefficient if
the SparseVector is in Mem::CUDA.

Now, the input first gets converted to Mem::Main, assembled and then
converted back to whatever memory architecture it originally was.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>