reimplemented matrix inversion from scratch
The previous implementation was zuguttenberged from the book 'Numerical
Recipes', which has a somewhat fishy license model that might have gotten
us into trouble some day in far future. Therefore, the matrix inversion
function has been reimplemented from scratch, featuring an enhanced
column-pivoted Gauss-Jordan elimination algorithm.

Although the previous algorithm featured total pivoting instead of
column-only pivoting, excessive testing with Hilbert, Lehmer and random
matrices did not reveal a significant loss in precision.

As a positive side-effect, the new implementation is roughly twice as fast
as the previous one and requires only 1*n instead of 3*n temporary memory.

Moreover, the invert_matrix function now also returns the determinant of
the input matrix, as this can be computed at virtually no additional cost,
because it is simply the product of all pivot elements. Due to this new
feature, the Tiny::Matrix template now also offers determinant computation
for dimensions > 6, which was not possible before.

Be aware that the determinant that is returned may in fact be infinity or
NaN, which simply means that either the determinant of the input matrix is
zero or that the input matrix already contained garbage entries.

This commit also adds the Math::isnormal function (see std::isnormal),
which checks whether a value is neither infinity, NaN, zero or subnormal.
This function can be used to check whether the result of the invert_matrix
function is 'useful' and therefore the matrix can be interpreted as
'successfully inverted'.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>