added support for 'fparser' 3rd party library
This commits adds support for the "Function Parser for C++" libary written
by Juha "Warp" Nieminen and Joel "Bisqwit" Yliluoma. Since the official
library name is somewhat lame and "non-unique", the library is simply
called 'fparser' in our context, which is also the corresponding build-id
tag that needs to be specified to include the library in your build.

This library provides a runtime function/formula/expression interpreter,
which is written in plain C++, i.e. it does not introduce any other
dependencies than the C++ standard library. This fact and its simplicity
are the library's main advantages in comparison to more powerful
alternatives such as e.g. python or lua. Furthermore, this library
provides a somewhat efficient implementation, as it compiles the parsed
expressions into custom bytecode, which is then "executed" for the
evaluation of the expressions.

The main purpose of this library in our context is to provide the means of
specifying various expressions as e.g. right-hand-side or boundary
condition functions at runtime via INI file or even via command line.
The corresponding formula (e.g. "2*x^3+y") can then be parsed and evaluated
by supplying various floating-point pairs for the two variables 'x' and 'y'.

Note:
The library's optimiser module had to be disabled if compiled with
Clang++, as there seems to be some problem with some SSE includes for this
compiler.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>