ENH: Added neighbour information to ConformalMesh
This is really easy to compute from the redundant index sets.

This is not implemented for StructuredMesh because saving neighbour
relations would be nonsensical and computing these relations on the fly
(like in the StructuredIndexMapping) is much slower due to boundary
identification. Maybe someone can come up with a way to bitshift that
whole if-if-else-if-else-if business.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>