ENH: Inverse mapping in Trafo
The inverse mapping is now implemented for the Standard transformation
on Simplex<d> meshes. Because the Q1 transformation is nonlinear, the
inverse mapping does not exist for Hypercube<d> meshes yet, so this was
not added to the Evaluator but exists as a set of seperate functions.
When someone implements the whole functionality later on, it can be
moved, but at the moment it is not clear that this will ever be used in
an assembly.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>