added Tutorial 04: parser
This new tutorial demonstrates the usage of the SimpleArgParser class,
which can be used to parse command line arguments. Moreover, if configured
and build with support for the 'fparser' third-party library, the tutorial
also demonstates the usage of the ParsedFunction class.

In combination, this tutorial combines the functionality of both parser
classes by implementing a simple Poisson solver (in analogy to tutorial
01), for which the user may specify the right-hand-side, boundary
condition and/or reference solution function at runtime via the command
line.

Note:
If configured and built without the 'fparser' library, this tutorial
provides a fallback, which uses the same function classes as tutorial 01.
In consequence, even without this additional library, the new tutorial
still does something useful, although the really interesting features
(runtime-specified functions) are not available, of course. The drawback
of this fallback is that the tutorial contains a few #ifdef blocks.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>