added String::parse(String&) overload
We already have a corresponding overload for our String class's base-class
std::string. However, some compilers chose to use the generic template
implementation instead of the std::string overload, which leads to
undesired behaviour.

This bug was found when trying to parse a command line argument containing
white-spaces enclosed in quotation marks, e.g.

> ./my_app "foo bar"

Without this new overload, trying to parse the argument via a construct
like

<code>
 String s;
 String(argv[1]).parse(s);
</code>

results in s="foo" rather than of s="foo bar".

The reason for this misbehaviour is that the generic template
implementation of String::parse uses the extraction operator >> of the
std::stringstream, which (by C++ standard) extracts only the first word
rather than the whole string.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>