ENH: Implemented DiscreteChart::project
This is somewhat ugly code as I wrote this in a very imperative manner
before and I am not that good at rethinking code if it works. But it
implements an orthogonal projection of a collection of points
indentified by a MeshPart to a surface mesh consisting of simplices.

Note that the meshes for DiscreteCharts HAVE to be of simplex shape.
This is because the whole
finding-surface-elements-to-project-points-to-them business requires an
inverse transformation (and the Q1 trafo is not linear, making this more
difficult) AND barycentric coordinates make a clever, adjacency based
search easy to implement.

The MeshPart to be projected is treated cell-wise and for each cell we
save information about where in the surface mesh the corresponding
vertices were found. So in practice, after the first cell was treated,
for every new vertex the complexity of finding the corresponding cell in
the surface mesh is O(1).

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>