simplified IterativeSolver class template interface
Previously, the IterativeSolver and PreconditionedIterativeSolver
base-classes were templatised in both the Matrix and Filter types.
This was necessary in an early draft of the solver design, but has
become widely redundant in the current design.

Due to this fact, the (Preconditioned)IterativeSolver class templates
are now templatised only in the underlying vector type. This simplifies
the usage when an actual iterative solver is created at runtime and
just used via its base-class pointer or reference.

This commit contains mostly 'under-the-hood' changes, so one only has
to adapt code if it uses the IterativeSolver interface directly, e.g.

  Solver::IterativeSolver<Matrix_, Filter_> solver = ...;

has to be replaced by

  Solver::IterativeSolver<typename Matrix_::VectorTypeR> solver = ...;

or
  Solver::IterativeSolver<Vector_> solver = ...;

if the type 'Vector_' specifies the corresponding vector type.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>