revised ScalarErrorComputer assembly class
The computation of H0-, H1- and/or H2- has been merged into a single class
template 'ScalarErrorComputer', which is templatised in the maximum
deriative order of the norm in which the error is to be measured, i.e.
the one and only template parameter specifies:

* = 0: compute only H0-error (aka L2-error)
* = 1: compute both H0- and H1-errors
* = 2: compute H0-, H1- and H2-errors

The return type of the error computing function is an object of the new
class 'ScalarErrorInfo', which holds all computed errors.

The primary reason for this change is performance: Computing both the H0-
(aka L2) error and the H1-error required two separate assembly sweeps
whereas only now a single sweep is necessary.

Note:
The console output of the applications using the scalar error computer now
print 'H0-Error' instead of 'L2-Error' for consistency. Note that by
definition 'H0' is the same as 'L2', so the error values are not affected.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>