implemented UmfpackMean solver
This commit introduces a new solver implementation named UmfpackMean,
which is capable of solving a linear system with the MeanFilter or any
other constraint that can be expressed as a scalar Lagrange multiplier for
the original system.

Technically, this class assembles the 'extended' saddle-point system
matrix including the Lagrange multiplier and then uses the Umfpack solver
for the factorisation and solution of the extended system.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>