implemented AnalyticVertex/CellProjector
These two assembly classes allow to interpolate an analytic function
directly in the vertices or cells of a mesh without taking a detour over
finite element spaces. Note that these two classes support the projection
of both scalar functions and vector fields. In the latter case, the
corresponding vector type is a DenseVectorBlocked.

Moreover, did some cleanup and minor refactoring of the Discrete*Projector
classes, thus also allowing the projection of a vector-valued finite element
function into a DenseVectorBlocked.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>