adjusted SparseMatrixBCSR::VectorTypeL/R
In the case there the block-width/height of a SparseMatrixBCSR is equal to
1, the corresponding right/left compatible vector type is now defined as a
simple DenseVector rather than a DenseVectorBlocked<...,1>.
The previous definition may have been more consistent, however, this would
lead to problems in meta-systems in which both vector-values and scalar
functions are discretised. One prominent example is the matrix of a simple
Stokes system, where the B and D matrices, which couple the velocity and
the pressure, are d-by-1 and 1-by-d BCSR-Matrices.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>