add enumerated type "Perspective"
This enumerated type is necessary to specify the treatment of
blocked datatypes like
- SparseMatrixBCSR
- SparseVectorBlocked
- DenseVectorBlocked
whether each block is treated as one entry (native perspective) or
each entry of a block is treated as its own (raw/pod perspective).

The member functions
- size()
- columns()
- rows()
- used_elements()
- val()
- elements()
of LAFEM containers are now using this enumerated type as an
template parameter to specify the perspective.
The former behaviour of these functions is still the same due to
the default template parameter Perspective::native.

Signed-off-by: Christoph Lohmann <christoph.lohmann@math.tu-dortmund.de>