ENH: Added StrongWolfeLinesearch
This is by far the most sophisticated linesearch implemented so far and
is a rewrite of of the linmin_mcstep and mcsrch routines that ALGLIB
inherited from MINPACK. The original implementation still contains more
black magick, but this is more readable and maintainable. If the
performance of the new implementation is significantly lower for real
applications (it is exactly the same for toy problems based on the
BazaraaShetty, Himmelblau and Rosenbrock functions), we can always do
more obscure vodoo later.

On termination, the linesearch guarantees that the strong Wolfe
conditions hold, which is a prerequisite for the convergence of the
nonlinear CG. If the strong Wolfe conditions do not hold, the linesearch
uses this knowledge for updating the interval of uncertainity with a
tricky polynomial fit.

The SecantLinesearch and NewtonRaphsonLinesearch are primitive compared
to this and need as much as a factor of ten (yes, 10!) more function and
gradient evaluations as they do not stop based on the strong Wolfe
conditions and thus need be as exact as possible, which requires many
iterations.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>