Removed moved statements, that prevented copy elision
From http://reviews.llvm.org/rL236075:
-Wpessimizing-move warns when a call to std::move would prevent copy elision
if the argument was not wrapped in a call. This happens when moving a local
variable in a return statement when the variable is the same type as the
return type or using a move to create a new object from a temporary object.

-Wredundant-move warns when an implicit move would already be made, so the
std::move call is not needed, such as when moving a local variable in a return
that is different from the return type.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>