ENH: Partial Meshopt rewrite, added MeshoptControl
Lots of changes in the Meshopt subsystem.

All mesh optimiser are now classified as mesh quality functionals named
by the underlying PDE measuring the mesh quality and are used in the new
MeshoptControl classes. This means:
  MeshSmoother -> MeshqualityFunctional
  LaplaceSmoother -> LaplaceFunctional
  DuDvSmoother -> DuDvFunctional
  RumpfSmoother*** -> HyperelasticityFunctional
  H_Evaluator -> RumpfTrafo

The (cell-) local functionals in the HyperelasticityFunctional are still
called RumpfFunctionals, because they use Rumpf's reference cells for
measuring deformation energy. The H_Evaluator, which evaluates
properties of that transformation has been renamed RumpfTrafo
accordingly.

All Q1Hack functionality has been put into the Q1Split (cell-) local
functional, finally ridding of of special classes. If such a functional
is created on a simplical mesh, it is just the default implementation.

Not all RumpfSmoother classes have been adapted to the new interface:
RumpfSmootherLevelset and RumpfSmootherLevelsetConc are dead code that
is kept for migration purposes only.

Even the nonlinear functionals now use classes from Solver instead of
their own implementations, which can be created dynamically with the
MeshoptSolverFactory. For the control classes, there is the
MeshoptControlFactory, which creates control objects at runtime based on
PropertyMaps. The most robust solvers are still ALGLIBMinCG and
ALGLIBMinLBFGS because they use the greatest amount of magic tricks, but
our own implementation in NLCG, NLSD and Linesearch is working fair
enough, and it's parallel.

There is a preconditioner for the nonlinear solvers called DuDv precon,
which assembles a D(u):D(v) matrix and applies a solver to it. This
works very well for problems with pure Dirichlet boundary conditions and
is not recommended for anything else. The implementation is very crude
and inefficient memory wise at the moment and subject to change.

Applications have been moved to their own subfolders because they come
with their own set of configuration .ini files. At the moment these
files just sit in the source folder and have to manually copied to the
build folder, or any working directory the corresponding application is
run from.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>