added RootMeshNode::extract_patch()
This new member function of the Geometry::RootMeshNode class implements
the functionality to extract a patch/partition of a Geometry mesh node
representing the base mesh into a new mesh node representing a single
patch/partition. Most of the actual splitting algorithms are implemented
and are ready to go.

The input required by this function is an adjacency graph representing the
ranks that are active on each element (thus also allowing non-trivial
overlap in paritioning), the rank of the patch to be extracted as well as
a vector of all ranks that are adjacent to the current rank.

The current implementations has a few limitations that will be lifted
later:

1. The mesh node may not have mesh parts which depend on other mesh parts,
i.e. the root mesh may not have "grand-child" mesh parts. Currently, we
have no mesh files which provide such mesh trees anyway.

2. Mesh part topologies and attributes are currently not partitioned, but
simply dropped. This is a serious limitation, as most of our mesh files
have explicit boundary parameterisation, however, no current application
requires this data. For our simple parallel applications, the current
implementation will do its job.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>
Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>