ENH: Default parameters as static constexpr
Linesearches and nonlinear optimisers had some parameters set to default
values in their constructors. And again by their new_***() function. And
again in the (soon to be) factory. If there is a sensible default for a
parameter, the class now constains it as a static constexpr, so the
factory and other routines can make use of it.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>