ContainerMainWrapper
This wrapper can serve as a represantative of any LAFEM::Container
object in main memory.
Thus a function, not yet ready for cuda support, can create a
ContainerMainWrapper instance from any lafem container and work on it as if the
original object resided in main memory.
Upon destruction, the wrapper object transfers all changes back to the
original memory.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>