added LAFEM::PowerXYZ::get() functions
This commits adds the get() member functions to all LAFEM::Power
meta-containers. The get() function is simply a runtime variant of the
already existing at() function, which allows to iterate over the
sub-blocks of a power container at runtime without the need of using
wicked template recursion magic. This should make life significantly
easier when trying to implement dimension-independent code, e.g. in case
of (Navier-)Stokes systems.

Be aware that this new get() function is only available for Power
containers, whereas its compile-time counterpart at() is available for all
meta container types. The reason is that the Power containers are the only
meta container type, where all blocks are always of the same type and thus
the return type of the function is independent of the block that is
requested.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>